# quadra-vite/src/api/generated/models/ResultMapStringLong.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ResultMapStringLong = {
    success?: boolean;
    code?: number;
    message?: string;
    data?: Record<string, number>;
    timestamp?: number;
    requestId?: string;
};